pyproject.toml
[project]
name = "mas-cdm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "pydantic-ai>=1.18.0",
    "pydantic-graph>=1.18.0",
    "redis>=7.1.0",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
